+++ types/protocol-mapping.d.ts
@@ -2239,13 +2239,6 @@ export namespace ProtocolMapping {
             returnType: void;
         };
         /**
-         * Emulates the given vision deficiency.
-         */
-        'Emulation.setEmulatedVisionDeficiency': {
-            paramsType: [Protocol.Emulation.SetEmulatedVisionDeficiencyRequest];
-            returnType: void;
-        };
-        /**
          * Overrides the Geolocation Position or Error. Omitting any of the parameters emulates position
          * unavailable.
          */